move client.js to a module; needs massive clean up; but there is a very basic test
