+++ bower.json
@@ -4,7 +4,6 @@
   "dependencies": {
     "angularjs": "1.3.0-beta.19",
     "bootstrap": "~3.2.0",
-    "d3": "~3.5.5",
     "jquery": "2.1.0",
     "jQuery-Storage-API": "~1.7.2",
     "jsSHA": "~1.5.0",
