+++ package.json
@@ -69,13 +69,15 @@
     "pushover-notifications": "0.2.0",
     "request": "^2.58.0",
     "sgvdata": "git://github.com/ktind/sgvdata.git#wip/protobuf",
-    "socket.io": "^1.3.5"
+    "socket.io": "^1.3.5",
+    "d3": "^3.5.6"
   },
   "devDependencies": {
     "istanbul": "~0.3.5",
     "mocha": "~1.20.1",
     "should": "~4.0.4",
     "supertest": "~0.13.0",
-    "jsdom": "^3.1.2"
+    "jsdom": "^3.1.2",
+    "benv": "^1.1.0"
   }
 }
